blog writing bot/requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.148
botocore==1.34.148
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.6.2
coloredlogs==15.0.1
crewai==0.41.1
crewai-tools==0.4.26
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
email_validator==2.2.0
embedchain==0.1.118
eval_type_backport==0.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastavro==1.9.5
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gptcache==0.1.43
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
iniconfig==2.0.0
instructor==1.3.3
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.11
langchain-cohere==0.1.9
langchain-community==0.2.10
langchain-core==0.2.23
langchain-experimental==0.0.63
langchain-groq==0.1.6
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
llama-cloud==0.0.11
llama-index==0.10.58
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.58
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.11
llama-index-embeddings-voyageai==0.1.4
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-groq==0.1.4
llama-index-llms-openai==0.1.27
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.31
llama-index-readers-llama-parse==0.1.6
llama-index-tools-wolfram-alpha==0.1.3
llama-parse==0.4.9
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mem0ai==0.0.5
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.37.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==1.5.1
parameterized==0.9.0
pillow==10.4.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.5.0
proto-plus==1.24.0
protobuf==4.25.4
pulsar-client==3.5.0
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pylance==0.9.18
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.373
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.10.1
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.23.1
semver==3.0.2
sentence-transformers==3.0.1
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
streamlit==1.31.0
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
together==1.2.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.4
transformers==4.43.3
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
validators==0.33.0
voyageai==0.1.7
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
